Add isDir and isFile tests
