+++ test/unit/editor/src/util/fileSystems/FsaEditorFileSystem/misc.test.js
@@ -23,94 +23,6 @@ Deno.test({
 });
 
 Deno.test({
-	name: "isFile true",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isFile = await fs.isFile(["root", "file1"]);
-
-		assertEquals(isFile, true);
-	},
-});
-
-Deno.test({
-	name: "isFile false",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isFile = await fs.isFile(["root"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isFile non existent",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isFile = await fs.isFile(["root", "nonExistent"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isFile non existent parent",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isFile = await fs.isFile(["root", "nonExistent", "file"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isDir true",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isDir = await fs.isDir(["root"]);
-
-		assertEquals(isDir, true);
-	},
-});
-
-Deno.test({
-	name: "isDir false",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isDir = await fs.isDir(["root", "file1"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
-	name: "isDir non existent",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isDir = await fs.isDir(["root", "nonExistent"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
-	name: "isDir non existent parent",
-	fn: async () => {
-		const {fs} = createBasicFs();
-
-		const isDir = await fs.isDir(["root", "nonExistent", "dir"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
 	name: "setRootName should throw",
 	async fn() {
 		const {fs} = createBasicFs();
