+++ test/unit/editor/src/util/fileSystems/IndexedDbEditorFileSystem/misc.test.js
@@ -257,116 +257,3 @@ Deno.test({
 		assertEquals(await promise2, "hello");
 	},
 });
-
-Deno.test({
-	name: "isFile true",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isFile = await fs.isFile(["root", "file1"]);
-
-		assertEquals(isFile, true);
-	},
-});
-
-Deno.test({
-	name: "isFile false",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isFile = await fs.isFile(["root"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isFile non existent",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isFile = await fs.isFile(["root", "nonExistent"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isFile non existent parent",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isFile = await fs.isFile(["root", "nonExistent", "file"]);
-
-		assertEquals(isFile, false);
-	},
-});
-
-Deno.test({
-	name: "isFile while it is being created",
-	async fn() {
-		const {fs} = await createBasicFs();
-
-		const promise1 = fs.writeFile(["root", "file"], "hello");
-		const promise2 = fs.isFile(["root", "file"]);
-		await promise1;
-		assertEquals(await promise2, true);
-	},
-});
-
-Deno.test({
-	name: "isDir true",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isDir = await fs.isDir(["root"]);
-
-		assertEquals(isDir, true);
-	},
-});
-
-Deno.test({
-	name: "isDir false",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isDir = await fs.isDir(["root", "file1"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
-	name: "isDir non existent",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isDir = await fs.isDir(["root", "nonExistent"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
-	name: "isDir non existent parent",
-	fn: async () => {
-		const {fs} = await createBasicFs();
-
-		const isDir = await fs.isDir(["root", "nonExistent", "dir"]);
-
-		assertEquals(isDir, false);
-	},
-});
-
-Deno.test({
-	name: "isDir while it is being created",
-	async fn() {
-		const {fs} = await createBasicFs();
-
-		const promise1 = fs.createDir(["root", "dir"]);
-		const promise2 = fs.isDir(["root", "dir"]);
-
-		await promise1;
-		assertEquals(await promise2, true);
-	},
-});
